requirements.txt
asgiref==3.2.7            # via django
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
django==3.0.5             # via -r djangorestframework
idna==2.9                 # via request
pyparsing==2.4.7          # via packaging
pytz==2019.3              # via django
requests==2.23.0
six==1.14.0               # via packaging
sqlparse==0.3.1           # via django
urllib3==1.25.9           # via requests
zipp==3.1.0               # via importlib-metadata
django-cors-headers==3.6.0
flake8==3.8.4
django-filter
gunicorn==20.0.4
pillow
psycopg2-binary
django-simple-email-confirmation==0.70
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.6.0
djangorestframework==3.11.0
sorl-thumbnail
